requirements.txt
python-dotenv
slack-bolt
google-genai